Implement `!rcsreload ccs` support.
RCS's `!rcsreload` command can be used by cohosts and up to force
reload room settings. This is useful eg. after they've just been
updated. This patch adds support for `!rcsreload ccs` to plug³.
When a chat message by a cohost or host comes in, that starts with
`!rcsreload ccs`, the room settings are reloaded.